LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__host.document-cookie.https-expected.txt


PASS __Host: Secure origin: Does not set 'Path=/;'
FAIL __HoSt: Secure origin: Does not set 'Path=/;' assert_equals: `__HoSt-prefixtestcookie=0.4276773007046819` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does set 'Secure; Path=/;'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; '
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; ' assert_equals: `__HoSt-prefixtestcookie=0.49744258705186106` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does not set 'Path=/;MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: `__HoSt-prefixtestcookie=0.7588813565710631` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=localhost; MaxAge=10' assert_equals: `__HoSt-prefixtestcookie=0.010830793442805242` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py' assert_equals: expected (undefined) undefined but got (string) "0.34217418780874154"
